Orchestrator/requirements.txt
# Orchestrator Conversational System - Requirements
# Core dependencies for the conversational AI orchestrator

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# HTTP Client
httpx>=0.25.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Environment Configuration
python-dotenv>=1.0.0

# Async Support
asyncio-mqtt>=0.16.0

# Logging
structlog>=23.2.0

# Optional: For enhanced audio processing
# numpy>=1.24.0
# scipy>=1.11.0

# Optional: For monitoring and metrics
# prometheus-client>=0.19.0
# psutil>=5.9.0
